Move 'clone' func to ghclone module
